refactor: move serde to eventually
