+++ eventually-postgres/src/lib.rs
@@ -3,7 +3,6 @@
 #![warn(clippy::pedantic)]
 #![warn(missing_docs)]
 
-pub mod serde;
 pub mod store;
 
 pub static MIGRATIONS: sqlx::migrate::Migrator = sqlx::migrate!("./migrations");
