+++ eventually/Cargo.toml
@@ -11,11 +11,17 @@ repository = "https://github.com/get-eventually/eventually-rs"
 categories = ["rust-patterns", "web-programming", "asynchronous", "data-structures"]
 keywords = ["architecture", "ddd", "event-sourcing", "cqrs", "es"]
 
+[features]
+default = []
+serde-prost = ["dep:prost"]
+full = ["serde-prost"]
+
 [dependencies]
 async-trait = "0.1.57"
 futures = "0.3.24"
 serde = { version = "1.0.130", features = ["derive"] }
 thiserror = "1.0.30"
+prost = { version = "0.11.0", optional = true }
 
 [dev-dependencies]
 anyhow = "1.0.51"  # NOTE: this is only used for test components and assertions.
