+++ eventually/src/lib.rs
@@ -7,4 +7,5 @@ pub mod aggregate;
 pub mod command;
 pub mod event;
 pub mod message;
+pub mod serde;
 pub mod version;
