+++ examples/bank-accounting/src/main.rs
@@ -1,9 +1,10 @@
 use std::time::Duration;
 
 use anyhow::anyhow;
+use eventually::serde::prost::MessageSerde;
 use eventually_postgres::store::EventStore;
 
-use bank_accounting::{application, domain::BankAccountRepository, grpc, proto, serde};
+use bank_accounting::{application, domain::BankAccountRepository, grpc, proto};
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -16,7 +17,8 @@ async fn main() -> anyhow::Result<()> {
     let database_url = std::env::var("DATABASE_URL").expect("env var DATABASE_URL is required");
     let pool = sqlx::PgPool::connect(&database_url).await?;
 
-    let bank_account_event_store = EventStore::new(pool, serde::BankAccountEventSerde).await?;
+    let bank_account_event_serde = MessageSerde::<proto::Event>::default();
+    let bank_account_event_store = EventStore::new(pool, bank_account_event_serde).await?;
     let bank_account_repository = BankAccountRepository::from(bank_account_event_store.clone());
 
     let application_service = application::Service::from(bank_account_repository);
